[UserFeedbacks] Notify the user with an email when a record is created
