+++ app/helpers/application_helper.rb
@@ -70,29 +70,6 @@ module ApplicationHelper
     tag.li(link_to(text, url, id: "#{id}-link", **options), id: id, class: klass)
   end
 
-  def dtext_ragel(text, **)
-    parsed = DText.parse(text, **)
-    return raw "" if parsed.nil?
-    deferred_post_ids.merge(parsed[1]) if parsed[1].present?
-    raw parsed[0]
-  rescue DText::Error => e
-    raw ""
-  end
-
-  def format_text(text, **options)
-    # preserve the currrent inline behaviour
-    if options[:inline]
-      dtext_ragel(text, **options)
-    else
-      raw %(<div class="styled-dtext">#{dtext_ragel(text, **options)}</div>)
-    end
-  end
-
-  def format_plaintext(text)
-    return "" if text.nil?
-    strip_tags(format_text(text).gsub(/<[^>]+>/, " ")).squish
-  end
-
   def custom_form_for(object, *args, &)
     options = args.extract_options!
     simple_form_for(object, *(args << options.merge(builder: CustomFormBuilder)), &)
