+++ app/views/layouts/mailer.html.mjml
@@ -8,6 +8,7 @@
       <mj-class name="hint" color="#bbbbbb" font-size="12px" line-height="16px" />
 
       <mj-class name="button" border-radius="4px" background-color="#d3b54d" color="#000" font-weight="bold" />
+
     </mj-attributes>
     <mj-style>
       a {
@@ -18,6 +19,7 @@
         text-decoration: underline;
       }
     </mj-style>
+    <%= yield :head %>
   </mj-head>
 	<mj-body background-color="#020f23">
     <mj-section padding="16px">
